+++ common/src/main/java/io/github/flemmli97/runecraftory/common/entities/BaseMonster.java
@@ -18,6 +18,7 @@ import io.github.flemmli97.runecraftory.common.entities.ai.StayGoal;
 import io.github.flemmli97.runecraftory.common.entities.ai.TendCropsGoal;
 import io.github.flemmli97.runecraftory.common.items.consumables.ItemObjectX;
 import io.github.flemmli97.runecraftory.common.lib.LibConstants;
+import io.github.flemmli97.runecraftory.common.loot.LootCtxParameters;
 import io.github.flemmli97.runecraftory.common.network.S2CAttackDebug;
 import io.github.flemmli97.runecraftory.common.network.S2COpenCompanionGui;
 import io.github.flemmli97.runecraftory.common.registry.ModAttributes;
@@ -48,6 +49,8 @@ import net.minecraft.network.protocol.game.ClientboundUpdateAttributesPacket;
 import net.minecraft.network.syncher.EntityDataAccessor;
 import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
@@ -93,6 +96,10 @@ import net.minecraft.world.level.GameRules;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.ServerLevelAccessor;
 import net.minecraft.world.level.entity.EntityTypeTest;
+import net.minecraft.world.level.storage.loot.LootContext;
+import net.minecraft.world.level.storage.loot.LootTable;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParams;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.Nullable;
@@ -105,7 +112,6 @@ import java.util.Optional;
 import java.util.UUID;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
 public abstract class BaseMonster extends PathfinderMob implements Enemy, IAnimated, IExtendedMob, RandomAttackSelectorMob, ExtendedEntity {
 
@@ -193,7 +199,6 @@ public abstract class BaseMonster extends PathfinderMob implements Enemy, IAnima
     private Behaviour behaviour = Behaviour.WANDER;
 
     private final LevelExpPair friendlyPoints = new LevelExpPair();
-    private Map<ItemStack, Integer> dailyDrops;
 
     private final DailyMonsterUpdater updater = new DailyMonsterUpdater(this);
 
@@ -694,12 +699,9 @@ public abstract class BaseMonster extends PathfinderMob implements Enemy, IAnima
         return ModTags.tamingTag(this.getType());
     }
 
-    @Override
-    public Map<ItemStack, Integer> dailyDrops() {
-        if (this.dailyDrops == null)
-            this.dailyDrops = this.prop.dailyDrops().entrySet().stream()
-                    .collect(Collectors.toMap(e -> e.getKey().getStack(), Map.Entry::getValue));
-        return this.dailyDrops;
+    public ResourceLocation dailyDropTable() {
+        ResourceLocation def = this.getDefaultLootTable();
+        return new ResourceLocation(def.getNamespace(), def.getPath() + "_tamed_drops");
     }
 
     @Override
@@ -939,13 +941,30 @@ public abstract class BaseMonster extends PathfinderMob implements Enemy, IAnima
     }
 
     @Override
-    public int friendPoints(Player player) {
-        if (player.getUUID().equals(this.getOwnerUUID()))
+    public int friendPoints(UUID player) {
+        if (player.equals(this.getOwnerUUID()))
             return this.entityData.get(friendPointsSync);
         return 0;
     }
 
     public void onDailyUpdate() {
+        if (this.level instanceof ServerLevel && this.isTamed()) {
+            ResourceLocation resourceLocation = this.dailyDropTable();
+            this.dropAsDailyDrop(resourceLocation);
+        }
+    }
+
+    protected void dropAsDailyDrop(ResourceLocation resourceLocation) {
+        LootTable lootTable = this.level.getServer().getLootTables().get(resourceLocation);
+        lootTable.getRandomItems(this.dailyDropContext().create(LootContextParamSets.GIFT), this::spawnAtLocation);
+    }
+
+    protected LootContext.Builder dailyDropContext() {
+        return new LootContext.Builder((ServerLevel) this.level)
+                .withRandom(this.random)
+                .withParameter(LootContextParams.THIS_ENTITY, this)
+                .withParameter(LootContextParams.ORIGIN, this.position())
+                .withParameter(LootCtxParameters.UUID_CONTEXT, this.getOwnerUUID());
     }
 
     //=====Damage Logic