+++ common/src/main/java/io/github/flemmli97/runecraftory/common/entities/IBaseMob.java
@@ -4,13 +4,19 @@ import io.github.flemmli97.runecraftory.common.attachment.player.LevelExpPair;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 
+import java.util.UUID;
+
 public interface IBaseMob {
 
     LevelExpPair level();
 
     void setLevel(int level);
 
-    int friendPoints(Player player);
+    default int friendPoints(Player player) {
+        return this.friendPoints(player.getUUID());
+    }
+
+    int friendPoints(UUID uuid);
 
     int baseXP();